Fixed bugs: exit to main menu and load new game works now. Putting a post-it where an opponent put a post-it no longer removes opponent's post-it

+++ UnityProject/Assets/HUD/PostIt.cs
@@ -33,7 +33,7 @@ public class PostIt : MonoBehaviour {
         VisibilityController.instance.VisibilityChanged += visHandler;
 
         positionHandler = (NodeData node, PostIt claimant) => {
-            if (node == currentNode && claimant != this) {
+            if (node == currentNode && TurnController.instance.CurrentPlayer == owner && claimant != this) {
                 destroySelf();
             }
         };
